add test by properties tutorial

+++ operations/testing-tutorial/properties-test-tutorial/src/main/java/fr/eletutour/properties/testing/ScoreTools.java
@@ -2,9 +2,7 @@ package fr.eletutour.properties.testing;
 
 public class ScoreTools {
 
-    public static int applyBonus(int score,
-                                 int bonusPct) {
-
+    public static long applyBonus(long score, int bonusPct) {
         return score + (score * bonusPct / 100);
     }
 }
